Fix last lint errors
