+++ lib/handlers.js
@@ -79,14 +79,14 @@ module.exports.putDocState = function (req, res) {
   documentService.setDocumentState(req.resource, req.filter, req.body.from, req.body.to, req.user)
     .then((result) => {
       return helpers.json(res, result);
-    }).catch((kind, error) => {
-      switch (kind) {
-        case 'validation':
+    }).catch((err) => {
+      switch (err.message) {
+        case 'Validation Error':
           return helpers.validationError(res);
-        case 'notFound':
+        case 'Not Found':
           return helpers.notFound(res);
         default:
-          return helpers.error(res, error);
+          return helpers.error(res, err);
       }
     });
 };
@@ -96,16 +96,16 @@ module.exports.putDoc = function (req, res) {
   documentService.setDocument(req.resource, req.filter, req.body, req.state, req.user)
     .then((result) => {
       return helpers.json(res, result);
-    }).catch((kind, error) => {
-      switch (kind) {
-        case 'validation': {
+    }).catch((err) => {
+      switch (err.message) {
+        case 'Validation Error': {
           let func = helpers.validationError(res);
-          return func(error);
+          return func(err);
         }
-        case 'notFound':
+        case 'Not Found':
           return helpers.notFound(res);
         default:
-          return helpers.error(res, error);
+          return helpers.error(res, err);
       }
     });
 };
@@ -126,12 +126,9 @@ module.exports.delDoc = function (req, res) {
     .then(() => {
       return helpers.json(res, {'message': 'OK'});
     })
-    .catch((error) => {
-      if (error) {
-        return helpers.error(res, error);
-      } else {
-        return helpers.notFound(res);
-      }
+    .catch((err) => {
+      if (err.message === 'Not Found') return helpers.notFound(res);
+      return helpers.error(res, err);
     });
 };
