+++ lib/modules/embedDocs.js
@@ -5,7 +5,7 @@ const builder = require('./queryBuilder');
 
 function fetchSubdoc (resource, reference, user, hash) {
   const _id = reference.get();
-  return new Promise((resolve) => {
+  return new Promise((resolve, reject) => {
     if (typeof hash[_id] !== 'undefined') {
       reference.set(hash[_id]);
       return resolve();
@@ -15,6 +15,7 @@ function fetchSubdoc (resource, reference, user, hash) {
       {_id: new ObjectId(_id)},
       builder.select({resource: resource, user: user}),
       (err, subDoc) => {
+        if (err) return reject(err);
         hash[_id] = subDoc.states[resource.defaultState].data;
         reference.set(hash[_id]);
         return resolve();
@@ -52,7 +53,7 @@ function embedDocs (resource, schema, embed, user, doc, hash) {
           hash
         ).then(refCb);
       }, cb);
-    }, () => (error) ? reject() : resolve());
+    }, () => (error) ? reject(error) : resolve());
   });
 }
 module.exports.one = embedDocs;
