+++ lib/param.js
@@ -25,7 +25,7 @@ function can (user, resource, method, state) {
       }
     }
 
-    // Test Permission fo owner role
+    // Test Permission for owner role
     if (resource.permissions['owner'] !== undefined) {
       let permission = resource.permissions['owner'][state];
       if (permission && (permission.includes(method) || permission === '*')) {
@@ -33,7 +33,7 @@ function can (user, resource, method, state) {
       }
     }
 
-    return reject();
+    return reject(new Error('unauthorized user'));
   });
 }
