+++ lib/services/document.js
@@ -67,6 +67,7 @@ module.exports.createDocument = function (resource, data, state, user) {
       user: user
     }).then((doc) => {
       resource.collection.insert(doc, (err, result) => {
+        if (err) throw err;
         resolve(
           Object.assign({
             state: state,
@@ -88,13 +89,11 @@ module.exports.setDocument = function (resource, filter, data, state, user) {
       state: state,
       user: user
     }).then((update) => {
-      resource.collection.updateOne(filter, update, (error, result) => {
-        if (error) {
-          return reject('generic', error);
-        }
+      resource.collection.updateOne(filter, update, (err, result) => {
+        if (err) return reject(err);
 
         if (result.modifiedCount !== 1) {
-          return reject('notFound');
+          return reject(new Error('Not Found'));
         }
         resolve({
           id: filter._id,
@@ -103,7 +102,7 @@ module.exports.setDocument = function (resource, filter, data, state, user) {
         });
       });
     }).catch(() => {
-      return reject('validation');
+      return reject(new Error('Validation Error'));
     });
   });
 };
@@ -124,11 +123,12 @@ module.exports.getDocument = function (resource, schema, filter, query, user, st
 
   return new Promise((resolve, reject) => {
     resource.collection.findOne(filter, fields, (err, doc) => {
+      if (err) return reject(err);
       if (doc === null) {
-        return reject();
+        return reject(new Error('Document Not Found'));
       }
       if (doc.states[state] === undefined) {
-        return reject();
+        return reject(new Error('Document Not Found'));
       }
 
       const currentState = doc.states[state] || {};
@@ -167,13 +167,11 @@ module.exports.setDocumentState = function (resource, filter, fromState, toState
         resource: resource,
         user: user
       }).then((ops) => {
-        resource.collection.updateOne(filter, ops, (error, result) => {
-          if (error) {
-            return reject('generic', error);
-          }
+        resource.collection.updateOne(filter, ops, (err, result) => {
+          if (err) return reject(err);
 
           if (result.modifiedCount !== 1) {
-            return reject('notFound');
+            return reject(new Error('Not Found'));
           }
 
           resolve({
@@ -184,25 +182,28 @@ module.exports.setDocumentState = function (resource, filter, fromState, toState
             }
           });
         });
-      }).catch(reject('validation'));
+      }).catch((err) => {
+        reject(err);
+      });
     };
 
     const stateTo = resource.states[toState];
     const stateFrom = resource.states[fromState];
 
     if (typeof stateTo === 'undefined') {
-      reject('generic', {message: 'Undefined state', state: toState});
+      reject(new Error(`Undefined state: ${toState}`));
     }
 
     if (typeof stateFrom === 'undefined') {
-      reject('generic', {message: 'Undefined state', state: fromState});
+      reject(new Error(`Undefined state: ${fromState}`));
     }
 
     if (!stateTo.validate) {
       return doSetState();
     }
 
-    resource.collection.findOne(filter, (error, document) => {
+    resource.collection.findOne(filter, (err, document) => {
+      if (err) return reject(err);
       doSetState(document.states[fromState].data);
     });
   });
@@ -216,8 +217,9 @@ module.exports.getDocumentState = function (resource, filter, state, user) {
   });
   return new Promise((resolve, reject) => {
     resource.collection.findOne(filter, fields, (err, doc) => {
+      if (err) reject(err);
       if (doc === null) {
-        return reject();
+        return reject(new Error('Not Found'));
       }
 
       resolve({
@@ -234,12 +236,8 @@ module.exports.deleteDocument = function (resource, filter, state) {
   updateParams.$unset[statePath] = '';
   return new Promise((resolve, reject) => {
     resource.collection.findOneAndUpdate(filter, updateParams, (err, out) => {
-      if (err) {
-        return reject(err);
-      }
-      if (out.lastErrorObject.n === 0) {
-        return reject();
-      }
+      if (err) return reject(err);
+      if (out.lastErrorObject.n === 0) return reject(new Error('Not Found'));
       let filter = builder.deleteFilter({id: out.value._id});
       resource.collection.findOneAndDelete(filter, () => {
         return resolve();
